requirements.txt
flask
flask-pymongo
requests
python-dotenv
pyjwt